Adding turn annoucement and player healths
