+++ apps/client/src/features/Game/GameObject/Character/Hero/Barbarian.tsx
