+++ apps/client/src/features/Game/GameObject/Character/Skeleton/Mage.tsx
@@ -107,15 +107,15 @@ interface GLTFAction extends THREE.AnimationClip {
 
 type GLTFResult = GLTF & {
   nodes: {
-    Skeleton_Minion_ArmLeft: THREE.SkinnedMesh
-    Skeleton_Minion_ArmRight: THREE.SkinnedMesh
-    Skeleton_Minion_Body: THREE.SkinnedMesh
-    Skeleton_Minion_Cloak: THREE.SkinnedMesh
-    Skeleton_Minion_Eyes: THREE.SkinnedMesh
-    Skeleton_Minion_Head: THREE.SkinnedMesh
-    Skeleton_Minion_Jaw: THREE.SkinnedMesh
-    Skeleton_Minion_LegLeft: THREE.SkinnedMesh
-    Skeleton_Minion_LegRight: THREE.SkinnedMesh
+    Skeleton_Mage_ArmLeft: THREE.SkinnedMesh
+    Skeleton_Mage_ArmRight: THREE.SkinnedMesh
+    Skeleton_Mage_Body: THREE.SkinnedMesh
+    Skeleton_Mage_Eyes: THREE.SkinnedMesh
+    Skeleton_Mage_Jaw: THREE.SkinnedMesh
+    Skeleton_Mage_LegLeft: THREE.SkinnedMesh
+    Skeleton_Mage_LegRight: THREE.SkinnedMesh
+    Skeleton_Mage_Skull: THREE.SkinnedMesh
+    Skeleton_Mage_Hat: THREE.Mesh
     root: THREE.Bone
   }
   materials: {
@@ -125,14 +125,14 @@ type GLTFResult = GLTF & {
   animations: GLTFAction[]
 }
 
-const MinionSkeleton: React.FC<JSX.IntrinsicElements['group']> = props => {
+const MageSkeleton: React.FC<JSX.IntrinsicElements['group']> = props => {
   const transform = React.useRef<THREE.Group>(null!)
-  const { scene, materials, animations } = useGLTF('/arts/skeletons/characters/gltf/Skeleton_Minion.glb') as unknown as GLTFResult
+  const { scene, materials, animations } = useGLTF('/arts/skeletons/characters/gltf/Skeleton_Mage.glb') as unknown as GLTFResult
   const clone = React.useMemo(() => SkeletonUtils.clone(scene), [scene])
   const { nodes } = useGraph(clone) as unknown as GLTFResult
 
   const { actions } = useAnimations<GLTFAction>(animations, transform)
-  const [animation] = React.useState<ActionName>('Idle_B')
+  const [animation] = React.useState<ActionName>('Idle_Combat')
 
   React.useEffect(() => {
     actions[animation]?.reset()?.fadeIn(0.1)?.play()
@@ -147,58 +147,52 @@ const MinionSkeleton: React.FC<JSX.IntrinsicElements['group']> = props => {
       <group name="Scene">
         <group name="Rig">
           <skinnedMesh
-            name="Skeleton_Minion_ArmLeft"
-            geometry={nodes.Skeleton_Minion_ArmLeft.geometry}
+            name="Skeleton_Mage_ArmLeft"
+            geometry={nodes.Skeleton_Mage_ArmLeft.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Minion_ArmLeft.skeleton}
+            skeleton={nodes.Skeleton_Mage_ArmLeft.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Minion_ArmRight"
-            geometry={nodes.Skeleton_Minion_ArmRight.geometry}
+            name="Skeleton_Mage_ArmRight"
+            geometry={nodes.Skeleton_Mage_ArmRight.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Minion_ArmRight.skeleton}
+            skeleton={nodes.Skeleton_Mage_ArmRight.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Minion_Body"
-            geometry={nodes.Skeleton_Minion_Body.geometry}
+            name="Skeleton_Mage_Body"
+            geometry={nodes.Skeleton_Mage_Body.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Minion_Body.skeleton}
+            skeleton={nodes.Skeleton_Mage_Body.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Minion_Cloak"
-            geometry={nodes.Skeleton_Minion_Cloak.geometry}
-            material={materials.skeleton}
-            skeleton={nodes.Skeleton_Minion_Cloak.skeleton}
-          />
-          <skinnedMesh
-            name="Skeleton_Minion_Eyes"
-            geometry={nodes.Skeleton_Minion_Eyes.geometry}
+            name="Skeleton_Mage_Eyes"
+            geometry={nodes.Skeleton_Mage_Eyes.geometry}
             material={materials.Glow}
-            skeleton={nodes.Skeleton_Minion_Eyes.skeleton}
+            skeleton={nodes.Skeleton_Mage_Eyes.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Minion_Head"
-            geometry={nodes.Skeleton_Minion_Head.geometry}
+            name="Skeleton_Mage_Jaw"
+            geometry={nodes.Skeleton_Mage_Jaw.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Minion_Head.skeleton}
+            skeleton={nodes.Skeleton_Mage_Jaw.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Minion_Jaw"
-            geometry={nodes.Skeleton_Minion_Jaw.geometry}
+            name="Skeleton_Mage_LegLeft"
+            geometry={nodes.Skeleton_Mage_LegLeft.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Minion_Jaw.skeleton}
+            skeleton={nodes.Skeleton_Mage_LegLeft.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Minion_LegLeft"
-            geometry={nodes.Skeleton_Minion_LegLeft.geometry}
+            name="Skeleton_Mage_LegRight"
+            geometry={nodes.Skeleton_Mage_LegRight.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Minion_LegLeft.skeleton}
+            skeleton={nodes.Skeleton_Mage_LegRight.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Minion_LegRight"
-            geometry={nodes.Skeleton_Minion_LegRight.geometry}
+            name="Skeleton_Mage_Skull"
+            geometry={nodes.Skeleton_Mage_Skull.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Minion_LegRight.skeleton}
+            skeleton={nodes.Skeleton_Mage_Skull.skeleton}
           />
           <primitive object={nodes.root} />
         </group>
@@ -207,6 +201,6 @@ const MinionSkeleton: React.FC<JSX.IntrinsicElements['group']> = props => {
   )
 }
 
-useGLTF.preload('/arts/skeletons/characters/gltf/Skeleton_Minion.glb')
+useGLTF.preload('/arts/skeletons/characters/gltf/Skeleton_Mage.glb')
 
-export default MinionSkeleton
+export default MageSkeleton
