+++ apps/client/src/features/Game/GameObject/Character/Skeleton/Minion.tsx
@@ -107,16 +107,15 @@ interface GLTFAction extends THREE.AnimationClip {
 
 type GLTFResult = GLTF & {
   nodes: {
-    Skeleton_Rogue_ArmLeft: THREE.SkinnedMesh
-    Skeleton_Rogue_ArmRight: THREE.SkinnedMesh
-    Skeleton_Rogue_Body: THREE.SkinnedMesh
-    Skeleton_Rogue_Eyes: THREE.SkinnedMesh
-    Skeleton_Rogue_Head: THREE.SkinnedMesh
-    Skeleton_Rogue_Jaw: THREE.SkinnedMesh
-    Skeleton_Rogue_LegLeft: THREE.SkinnedMesh
-    Skeleton_Rogue_LegRight: THREE.SkinnedMesh
-    Skeleton_Rogue_Hood: THREE.Mesh
-    Skeleton_Rogue_Cape: THREE.Mesh
+    Skeleton_Minion_ArmLeft: THREE.SkinnedMesh
+    Skeleton_Minion_ArmRight: THREE.SkinnedMesh
+    Skeleton_Minion_Body: THREE.SkinnedMesh
+    Skeleton_Minion_Cloak: THREE.SkinnedMesh
+    Skeleton_Minion_Eyes: THREE.SkinnedMesh
+    Skeleton_Minion_Head: THREE.SkinnedMesh
+    Skeleton_Minion_Jaw: THREE.SkinnedMesh
+    Skeleton_Minion_LegLeft: THREE.SkinnedMesh
+    Skeleton_Minion_LegRight: THREE.SkinnedMesh
     root: THREE.Bone
   }
   materials: {
@@ -126,14 +125,14 @@ type GLTFResult = GLTF & {
   animations: GLTFAction[]
 }
 
-const RogueSkeleton: React.FC<JSX.IntrinsicElements['group']> = props => {
+const MinionSkeleton: React.FC<JSX.IntrinsicElements['group']> = props => {
   const transform = React.useRef<THREE.Group>(null!)
-  const { scene, materials, animations } = useGLTF('/arts/skeletons/characters/gltf/Skeleton_Rogue.glb') as unknown as GLTFResult
+  const { scene, materials, animations } = useGLTF('/arts/skeletons/characters/gltf/Skeleton_Minion.glb') as unknown as GLTFResult
   const clone = React.useMemo(() => SkeletonUtils.clone(scene), [scene])
   const { nodes } = useGraph(clone) as unknown as GLTFResult
 
   const { actions } = useAnimations<GLTFAction>(animations, transform)
-  const [animation] = React.useState<ActionName>('Idle_B')
+  const [animation] = React.useState<ActionName>('Idle_Combat')
 
   React.useEffect(() => {
     actions[animation]?.reset()?.fadeIn(0.1)?.play()
@@ -148,52 +147,58 @@ const RogueSkeleton: React.FC<JSX.IntrinsicElements['group']> = props => {
       <group name="Scene">
         <group name="Rig">
           <skinnedMesh
-            name="Skeleton_Rogue_ArmLeft"
-            geometry={nodes.Skeleton_Rogue_ArmLeft.geometry}
+            name="Skeleton_Minion_ArmLeft"
+            geometry={nodes.Skeleton_Minion_ArmLeft.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Rogue_ArmLeft.skeleton}
+            skeleton={nodes.Skeleton_Minion_ArmLeft.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Rogue_ArmRight"
-            geometry={nodes.Skeleton_Rogue_ArmRight.geometry}
+            name="Skeleton_Minion_ArmRight"
+            geometry={nodes.Skeleton_Minion_ArmRight.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Rogue_ArmRight.skeleton}
+            skeleton={nodes.Skeleton_Minion_ArmRight.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Rogue_Body"
-            geometry={nodes.Skeleton_Rogue_Body.geometry}
+            name="Skeleton_Minion_Body"
+            geometry={nodes.Skeleton_Minion_Body.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Rogue_Body.skeleton}
+            skeleton={nodes.Skeleton_Minion_Body.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Rogue_Eyes"
-            geometry={nodes.Skeleton_Rogue_Eyes.geometry}
+            name="Skeleton_Minion_Cloak"
+            geometry={nodes.Skeleton_Minion_Cloak.geometry}
+            material={materials.skeleton}
+            skeleton={nodes.Skeleton_Minion_Cloak.skeleton}
+          />
+          <skinnedMesh
+            name="Skeleton_Minion_Eyes"
+            geometry={nodes.Skeleton_Minion_Eyes.geometry}
             material={materials.Glow}
-            skeleton={nodes.Skeleton_Rogue_Eyes.skeleton}
+            skeleton={nodes.Skeleton_Minion_Eyes.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Rogue_Head"
-            geometry={nodes.Skeleton_Rogue_Head.geometry}
+            name="Skeleton_Minion_Head"
+            geometry={nodes.Skeleton_Minion_Head.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Rogue_Head.skeleton}
+            skeleton={nodes.Skeleton_Minion_Head.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Rogue_Jaw"
-            geometry={nodes.Skeleton_Rogue_Jaw.geometry}
+            name="Skeleton_Minion_Jaw"
+            geometry={nodes.Skeleton_Minion_Jaw.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Rogue_Jaw.skeleton}
+            skeleton={nodes.Skeleton_Minion_Jaw.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Rogue_LegLeft"
-            geometry={nodes.Skeleton_Rogue_LegLeft.geometry}
+            name="Skeleton_Minion_LegLeft"
+            geometry={nodes.Skeleton_Minion_LegLeft.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Rogue_LegLeft.skeleton}
+            skeleton={nodes.Skeleton_Minion_LegLeft.skeleton}
           />
           <skinnedMesh
-            name="Skeleton_Rogue_LegRight"
-            geometry={nodes.Skeleton_Rogue_LegRight.geometry}
+            name="Skeleton_Minion_LegRight"
+            geometry={nodes.Skeleton_Minion_LegRight.geometry}
             material={materials.skeleton}
-            skeleton={nodes.Skeleton_Rogue_LegRight.skeleton}
+            skeleton={nodes.Skeleton_Minion_LegRight.skeleton}
           />
           <primitive object={nodes.root} />
         </group>
@@ -202,6 +207,6 @@ const RogueSkeleton: React.FC<JSX.IntrinsicElements['group']> = props => {
   )
 }
 
-useGLTF.preload('/arts/skeletons/characters/gltf/Skeleton_Rogue.glb')
+useGLTF.preload('/arts/skeletons/characters/gltf/Skeleton_Minion.glb')
 
-export default RogueSkeleton
+export default MinionSkeleton
