+++ apps/client/src/features/Game/GameObject/Character/Skeleton/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Vector3 } from 'three'
 import { SkeletonType } from 'types/enemy'
-import MageSkeleton from 'features/Game/Character/Skeleton/Mage'
-import MinionSkeleton from 'features/Game/Character/Skeleton/Minion'
-import RogueSkeleton from 'features/Game/Character/Skeleton/Rogue'
-import WarriorSkeleton from 'features/Game/Character/Skeleton/Warrior'
+import MageSkeleton from 'features/Game/GameObject/Character/Skeleton/Mage'
+import MinionSkeleton from 'features/Game/GameObject/Character/Skeleton/Minion'
+import RogueSkeleton from 'features/Game/GameObject/Character/Skeleton/Rogue'
+import WarriorSkeleton from 'features/Game/GameObject/Character/Skeleton/Warrior'
 
 type Props = JSX.IntrinsicElements['group'] & {
   readonly type: SkeletonType
