+++ apps/client/src/features/Game/GameObject/Prop/Chest.tsx
