+++ apps/client/src/features/Game/GameObject/Tile/Corridor.tsx
