+++ apps/client/src/features/Game/Scene/Game/HUD/Players.tsx
@@ -11,9 +11,11 @@ const Players: React.FC = () => {
       <div className='flex w-full flex-wrap py-4 gap-4'>
         {players.map(player =>
           <PlayerCard
+            key={player.id}
             username={player.username}
             avatar={`/img/hero/${player.hero.toString()}.png`}
             weapons={player.inventory.weapons}
+            health={player.health}
           />
         )}
       </div>
