+++ apps/client/src/widgets/HUD/PlayerCard.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import Weapon from 'widgets/HUD/Weapon'
+import { FaHeart } from 'react-icons/fa'
 
 type WeaponLite = {
   readonly icon: string
@@ -8,18 +9,24 @@ type WeaponLite = {
 
 type Props = {
   readonly username: string
+  readonly health: number
   readonly avatar: string
   readonly weapons: WeaponLite[]
 }
 
-const PlayerCard: React.FC<Props> = ({ username, avatar, weapons }) => {
+const PlayerCard: React.FC<Props> = ({ username, avatar, health, weapons }) => {
   return (
     <div className='card card-side p-2 bg-gradient-to-tr from-orange-100 to-orange-300'>
       <figure>
         <img src={avatar} alt='avatar' className='w-24 h-24' />
       </figure>
       <div className='card-body'>
-        <h2 className='card-title text-black'>{username}</h2>
+        <div className='flex items-center gap-2'>
+          <h2 className='card-title text-black'>{username}</h2>
+          <div className='flex gap-1 items-center'>
+            {Array.from({ length: health }, () => <FaHeart className='text-red-500' />)}
+          </div>
+        </div>
         <div className='flex flex-col gap-2 justify-center items-start'>
           <div className='flex flex-row gap-2 justify-center'>
             {weapons.map(
